// src/app/shared/models/products.model.ts
export class MedicineModel {
    productId: number;
    categoryId:number;
    companyId: number;
    hsnTypeId: number;
    name: string;
    shortName: string;
    packing: number;
    units: number;
    createdAt: string;
    createdBy: number;
    updatedAt: string;
    udatedBy: number;
    status: boolean;
}